Explicitly cast decimals(uint8) to uint256

+++ contracts/MarginCalculator.sol
@@ -87,7 +87,7 @@ contract MarginCalculator {
 
         FPI.FixedPointInt memory collateralAmount = ZERO;
         if (hasCollateral) {
-            uint256 colllateralDecimals = ERC20Interface(_vault.collateralAssets[0]).decimals();
+            uint256 colllateralDecimals = uint256(ERC20Interface(_vault.collateralAssets[0]).decimals());
             collateralAmount = FPI.fromScaledUint(_vault.collateralAmounts[0], colllateralDecimals);
         }
 
@@ -98,7 +98,7 @@ contract MarginCalculator {
         bool isExcess = excessCollateral.isGreaterThanOrEqual(ZERO);
 
         address otoken = hasLong ? _vault.longOtokens[0] : _vault.shortOtokens[0];
-        uint256 collateralDecimals = ERC20Interface(OtokenInterface(otoken).collateralAsset()).decimals();
+        uint256 collateralDecimals = uint256(ERC20Interface(OtokenInterface(otoken).collateralAsset()).decimals());
         // if is excess, truncate the tailing digits in excessCollateralExternal calculation
         uint256 excessCollateralExternal = excessCollateral.toScaledUint(collateralDecimals, isExcess);
         return (excessCollateralExternal, isExcess);
